Update 1_antenatal_care.do
Sorry, there was a mistake in the template I shared, where h42 is only for the child iron supplement taking, Only m45 should be considered as it specified the taking for mother during pregnancy. If it's missing for m45, for example, in Bangladesh cases, then the c_anc_ir should be taken as missing.